+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Types of changes:
 ### Fixed
 
 - Fix query wrapping
+- Fix date column type casting in SQL queries to handle TIMESTAMP values correctly
 
 ## [0.6.0] - 2026-01-07